refactor(SwapAndAdd): second-round review cleanups
- ISwapAndAdd route docs: "bounded Permit2 allowances" -> standing; the
  per-operation bounded approvals were replaced by standing max
  allowances in the _ensureApproved rewrite and the doc never followed.
- _run takes UnlockData directly, so every call site is a named-field
  struct literal instead of positional magic zeros.
- _growCore returns only the CoreParams (the PoolKey was a duplicate of
  cp.key; compound discarded it).
- _pullAdditional folded into _pullBudget: rebalance clamps its positive
  deltas to a plain budget pull, deleting the near-twin function and the
  duplicated msg.value/balance guards.
- InsufficientLiquidity(minLiquidity) widened to uint256: a floor above
  uint128.max always reverted correctly but reported a truncated value.
- UR_SWEEP_COMMAND declared uint8 (its only use).
- Contract header trimmed to the implementation-specific INVARIANT
  (now noting per-OP granularity under multicall) plus a pointer to
  ISwapAndAdd, which owns all user-facing semantics.

No behavior change: 791 tests green under --isolate with snapshot
check; bytecode 18,485 -> 18,079.

+++ src/interfaces/ISwapAndAdd.sol
@@ -101,7 +101,7 @@ interface ISwapAndAdd is IMulticall_v4 {
     error InvalidEthValue();
     /// @notice Plain ETH transfers are rejected; only the PoolManager, POSM and the Universal Router may send.
     error InvalidEthSender();
-    error InsufficientLiquidity(uint128 minLiquidity, uint128 liquidity);
+    error InsufficientLiquidity(uint256 minLiquidity, uint128 liquidity);
     /// @notice Output may not be sent to this contract (it would strand there) or to the zero address (the
     ///         NFT hand-off rejects it late; swept leftovers would burn silently on the grow ops).
     error InvalidRecipient(address recipient);
@@ -134,7 +134,7 @@ interface ISwapAndAdd is IMulticall_v4 {
     /// @param amount1In     Budget of pool token1 (may be 0).
     /// @param route         Verbatim Universal Router payload for the bulk swap: abi.encode(bytes commands, bytes[] inputs).
     ///                      May be empty -> the whole deficit is sourced by the same-pool swap + trim. The route
-    ///                      declares its own input token and amount; BOTH pool tokens are funded for it (bounded
+    ///                      declares its own input token and amount; BOTH pool tokens are funded for it (standing
     ///                      Permit2 allowances / native value), whatever it does not consume stays in the contract
     ///                      for the same-pool reconcile, and native value left in the router is reclaimed.
     ///                      ROUTE-CONSTRUCTION REQUIREMENT: the contract forwards its WHOLE native balance with
